requirements/development.txt
-r base.txt

# Development-specific packages
django-debug-toolbar==4.4.6
flake8==7.3.0
isort==7.0.0